+++ CHANGELOG.md
@@ -44,6 +44,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the `ProgressIndicator` component's label margin to be consistent with the VDL.
 - Updates the `"dl"` variant of the `List` component to use `2rem` for column spacing and to set the width of the `term` columnm to be a full `"250px"` for `tablet` and `desktop` viewports.
 - Updates the `TagSet` component to sync the border color styles with the VDL.
+- Updates the `MultiSelect` component to now allow 0 to be rendered for `itemCount`.
 
 ### Fixes