+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the `Tabs` component to sync the border color with the VDL.
 - Updates the `SubNav` component to add dark mode color styles.
 - Updates the `Heading` component to increase the font weight for the `size` styles.
+- Updates the Storybook docs for the `Tooltip` component to clarify how a tooltip should be used.
 
 ## 3.6.0 (April 10, 2025)