+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates the `SubNav` component to add dark mode color styles.
 - Updates the `Heading` component to increase the font weight for the `size` styles.
 - Updates the Storybook docs for the `Tooltip` component to clarify how a tooltip should be used.
+- Updates the Storybook docs for the `Tabs` component to improve details about the mobile carousel and horizontal scrolling.
 
 ## 3.6.0 (April 10, 2025)